Renamed `MapPosts` to `Forums`, `MapTag` to `Tag`, created a new `Post` object and `Comment` object
May God forgive me. This required nuking the database so many times. I also completely deleted the `migrations` and `__pycahce__` folders (whatever those do). There may still be some residual issues, which I apologize for!

I also added some more features.

Forums now have the `associated` option, for the creator of the Forum to decide if they are associated with the group providing the solution. And they also have the `private_public` option, for the creator to decide if the Forum is open to the public or invite only. Lastly, Forums have the `contributors` attribute to determine if a Contributor has been invited to the Forum.

Posts can have Media; Comments cannot.